backend/loader/pyproject.toml
[tool.poetry]
name = "sparrow-loader"
version = "1.0.8"
description = "Utilities for loading data into Sparrow"
authors = ["Daven Quinn <dev@davenquinn.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "sparrow"}]
include = ["sparrow/loader/sparrow-db-models.pickle"]

[tool.poetry.dependencies]
python = "^3.9"
marshmallow = "^3.11.1"
marshmallow-sqlalchemy = "^0.24.2"
GeoAlchemy2 = "^0.9.4"
SQLAlchemy = "^1.4.41"
"macrostrat.database" = "^1.0.2"
stringcase = "^1.2.0"
marshmallow-jsonschema = "^0.10.0"
shapely = "^1"
typer = "^0.6.1"
rich = "^12.6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"

# Command-line scripts
[tool.poetry.scripts]
sparrow-loader = "sparrow.loader.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
